+++ src/frontend/org/voltdb/sysprocs/Promote.java
@@ -61,7 +61,7 @@ public class Promote extends VoltSystemProcedure {
     {
         // Choose the lowest site ID on this host to actually flip the bit
         int host_id = ctx.getExecutionSite().getCorrespondingHostId();
-        Integer lowest_site_id =
+        Long lowest_site_id =
             VoltDB.instance().getCatalogContext().siteTracker.
             getLowestLiveExecSiteIdForHost(host_id);
         if (ctx.getExecutionSite().getSiteId() == lowest_site_id)
